requirements/dev.txt
aniso8601==6.0.0
Django==2.2.3
django-filter==2.2.0
django-graphql-jwt==0.2.1
graphene==2.1.7
graphene-django==2.4.0
graphql-core==2.2.1
graphql-relay==2.0.0
promise==2.2.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
